Add `ParserVariableRegistry` to manage parser functions (and other parser things in the future) separately from the hook handler.
